Final README file placement / configuration update.
This removes the file from the "src" folder and places it in the root
repository directory; this change will be propigated to GitHub through
"push to origin" command in Eclipse IDE. If all goes well, GitHub will
recognise this README file the project's ONLY one; at that point, I need
edit only the one README for the whole project. Later, I may look into
the possibility of separate README files for each branch, to document
how that branch differs from the overall project and its overarching
README file at the root of the repository.